+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 - Update to Manifest V3
+- Hide broken shortcuts link in Firefox
 
 ## 1.4.1 - 2022 Oct 17
 - Rebuild